refactor: 채팅방 입장/퇴장/정보 조회 로직 서비스 함수로 분리
